clean up resolver control flow
- don't do work in destructors
- expecially don't do work that can throw in destructors
- simplify the parser <=> composite API touchpoint; nix a header
