+++ multibody/parsing/parser.cc
@@ -8,7 +8,6 @@
 #include "drake/multibody/parsing/detail_common.h"
 #include "drake/multibody/parsing/detail_composite_parse.h"
 #include "drake/multibody/parsing/detail_instanced_name.h"
-#include "drake/multibody/parsing/detail_parser_internal_data.h"
 #include "drake/multibody/parsing/detail_parsing_workspace.h"
 #include "drake/multibody/parsing/detail_path_utils.h"
 #include "drake/multibody/parsing/detail_select_parser.h"
@@ -23,6 +22,18 @@ using internal::DataSource;
 using internal::ParserInterface;
 using internal::ParsingWorkspace;
 using internal::SelectParser;
+
+namespace internal {
+
+// Storage for internals that need to have the same lifetime as a Parser
+// instance, but avoid adding internal namespace details to parser.h.
+struct ParserInternalData {
+  // Collision filter groups that use InstancedName. This representation is
+  // invariant with respect to model renaming via the plant.
+  CollisionFilterGroupsImpl<InstancedName> collision_filter_groups_storage_;
+};
+
+}  // namespace internal
 
 Parser::Parser(MultibodyPlant<double>* plant,
                geometry::SceneGraph<double>* scene_graph)
@@ -90,8 +101,10 @@ std::vector<ModelInstanceIndex> Parser::AddModels(
   DataSource data_source(DataSource::kFilename, &filename_string);
   ParserInterface& parser = SelectParser(diagnostic_policy_, file_name);
   auto composite = internal::CompositeParse::MakeCompositeParse(this);
-  return parser.AddAllModels(data_source, model_name_prefix_,
-                             composite->workspace());
+  auto result = parser.AddAllModels(data_source, model_name_prefix_,
+                                    composite->workspace());
+  composite->Finish();
+  return result;
 }
 
 std::vector<ModelInstanceIndex> Parser::AddModelsFromUrl(
@@ -110,8 +123,36 @@ std::vector<ModelInstanceIndex> Parser::AddModelsFromString(
   const std::string pseudo_name(data_source.GetStem() + "." + file_type);
   ParserInterface& parser = SelectParser(diagnostic_policy_, pseudo_name);
   auto composite = internal::CompositeParse::MakeCompositeParse(this);
-  return parser.AddAllModels(data_source, model_name_prefix_,
-                             composite->workspace());
+  auto result = parser.AddAllModels(data_source, model_name_prefix_,
+                                    composite->workspace());
+  composite->Finish();
+  return result;
+}
+
+void Parser::ResolveCollisionFilterGroupsFromCompositeParse(
+    internal::CollisionFilterGroupResolver* resolver) {
+  DRAKE_DEMAND(resolver != nullptr);
+
+  resolver->Resolve(diagnostic_policy_);
+
+  // Convert scoped names into InstancedNames for storage in between parses.
+  auto convert = [this](const std::string& input)
+                 -> internal::InstancedName {
+    internal::InstancedName result;
+    auto scoped = ScopedName::Parse(input);
+    if (plant_->HasModelInstanceNamed(scoped.get_namespace())) {
+      result.index = plant_->GetModelInstanceByName(scoped.get_namespace());
+    }
+    result.name = scoped.get_element();
+    return result;
+  };
+
+  // Merge the groups found during a composite parse into the accumulated groups
+  // held by this parser.
+  MergeCollisionFilterGroups<internal::InstancedName, std::string>(
+      &data_->collision_filter_groups_storage_,
+      resolver->GetCollisionFilterGroups(),
+      convert);
 }
 
 }  // namespace multibody
